refacto: use Level const wverywhere and clean the module imports
