+++ lbcluster/lbcluster_dns.go
@@ -6,77 +6,106 @@ import (
 	"time"
 
 	"github.com/miekg/dns"
+
+	"lb-experts/golbd/log"
 )
 
-/*RefreshDNS This is the only public function here. It retrieves the current ips behind the dns,
-and then updates it with the new best ips (if they are different) */
-func (lbc *LBCluster) RefreshDNS(dnsManager, keyPrefix, internalKey, externalKey string) {
+// Logger struct for the Logger interface
+type Logger interface {
+	Info(s string) error
+	Warning(s string) error
+	Debug(s string) error
+	Error(s string) error
+}
 
-	e := lbc.GetStateDNS(dnsManager)
-	if e != nil {
-		lbc.WriteToLog("WARNING", fmt.Sprintf("Get_state_dns Error: %v", e.Error()))
+// WriteToLog puts something in the log file
+func (lbc *LBCluster) WriteToLog(level string, input string) error {
+	msg := fmt.Sprintf("cluster: %s, %s", lbc.ClusterName, input)
+
+	switch level {
+	case log.LevelInfo:
+		return lbc.Slog.Info(msg)
+	case log.LevelDebug:
+		return lbc.Slog.Debug(msg)
+	case log.LevelWarning:
+		return lbc.Slog.Warning(msg)
+	case log.LevelError:
+		return lbc.Slog.Error(msg)
+	default:
+		return lbc.Slog.Error(fmt.Sprintf("unsupported level %s, assuming error %s", input, msg))
+	}
+}
+
+// RefreshDNS This is the only public function here. It retrieves the current ips behind the dns,
+// and then updates it with the new best ips (if they are different).
+func (lbc *LBCluster) RefreshDNS(dnsManager, keyPrefix, internalKey, externalKey string) {
+	err := lbc.GetStateDNS(dnsManager)
+	if err != nil {
+		lbc.WriteToLog(log.LevelWarning, fmt.Sprintf("GetStateDNS Error: %v", err.Error()))
 	}
 
 	pbiDNS := lbc.concatenateIps(lbc.PreviousBestIpsDns)
 	cbi := lbc.concatenateIps(lbc.CurrentBestIps)
 	if pbiDNS == cbi {
-		lbc.WriteToLog("INFO", fmt.Sprintf("DNS not update keyName %v cbh == pbhDns == %v", keyPrefix, cbi))
+		lbc.WriteToLog(log.LevelInfo, fmt.Sprintf("DNS not update keyName %v cbh == pbhDns == %v", keyPrefix, cbi))
 		return
 	}
 
-	lbc.WriteToLog("INFO", fmt.Sprintf("Updating the DNS with %v (previous state was %v)", cbi, pbiDNS))
+	lbc.WriteToLog(log.LevelInfo, fmt.Sprintf("Updating the DNS with %v (previous state was %v)", cbi, pbiDNS))
 
-	e = lbc.updateDNS(keyPrefix+"internal.", internalKey, dnsManager)
-	if e != nil {
-		lbc.WriteToLog("WARNING", fmt.Sprintf("Internal Update_dns Error: %v", e.Error()))
+	err = lbc.updateDNS(keyPrefix+"internal.", internalKey, dnsManager)
+	if err != nil {
+		lbc.WriteToLog(log.LevelWarning, fmt.Sprintf("Internal updateDNS Error: %v", err.Error()))
 	}
 	if lbc.externallyVisible() {
-		e = lbc.updateDNS(keyPrefix+"external.", externalKey, dnsManager)
-		if e != nil {
-			lbc.WriteToLog("WARNING", fmt.Sprintf("External Update_dns Error: %v", e.Error()))
+		err = lbc.updateDNS(keyPrefix+"external.", externalKey, dnsManager)
+		if err != nil {
+			lbc.WriteToLog(log.LevelWarning, fmt.Sprintf("External updateDNS Error: %v", err.Error()))
 		}
 	}
 }
 
-//Internal functions
 func (lbc *LBCluster) externallyVisible() bool {
 	return lbc.Parameters.External
 }
 
-func (lbc *LBCluster) updateDNS(keyName, tsigKey, dnsManager string) error {
+const (
+	defaultTTL = 60
+)
 
-	ttl := "60"
-	if lbc.Parameters.Ttl > 60 {
-		ttl = fmt.Sprintf("%d", lbc.Parameters.Ttl)
+func (lbc *LBCluster) updateDNS(keyName, tsigKey, dnsManager string) error {
+	ttl := defaultTTL
+	if lbc.Parameters.Ttl > defaultTTL {
+		ttl = lbc.Parameters.Ttl
 	}
-	//best_hosts_len := len(lbc.Current_best_hosts)
+
 	m := new(dns.Msg)
 	m.SetUpdate(lbc.ClusterName + ".")
 	m.Id = 1234
-	rrRemoveA, _ := dns.NewRR(lbc.ClusterName + ". " + ttl + " IN A 127.0.0.1")
-	rrRemoveAAAA, _ := dns.NewRR(lbc.ClusterName + ". " + ttl + " IN AAAA ::1")
+	rrRemoveA, _ := dns.NewRR(fmt.Sprintf("%s. %d IN A 127.0.0.1", lbc.ClusterName, ttl))
+	rrRemoveAAAA, _ := dns.NewRR(fmt.Sprintf("%s. %d IN AAAA ::1", lbc.ClusterName, ttl))
 	m.RemoveRRset([]dns.RR{rrRemoveA})
 	m.RemoveRRset([]dns.RR{rrRemoveAAAA})
 
 	for _, ip := range lbc.CurrentBestIps {
 		var rrInsert dns.RR
 		if ip.To4() != nil {
-			rrInsert, _ = dns.NewRR(lbc.ClusterName + ". " + ttl + " IN A " + ip.String())
+			rrInsert, _ = dns.NewRR(fmt.Sprintf("%s. %d IN A %s", lbc.ClusterName, ttl, ip.String()))
 		} else if ip.To16() != nil {
-			rrInsert, _ = dns.NewRR(lbc.ClusterName + ". " + ttl + " IN AAAA " + ip.String())
+			rrInsert, _ = dns.NewRR(fmt.Sprintf("%s. %d IN IN AAAA %s", lbc.ClusterName, ttl, ip.String()))
 		}
 		m.Insert([]dns.RR{rrInsert})
 	}
-	lbc.WriteToLog("INFO", fmt.Sprintf("WE WOULD UPDATE THE DNS WITH THE IPS %v", m))
+	lbc.WriteToLog(log.LevelInfo, fmt.Sprintf("WE WOULD UPDATE THE DNS WITH THE IPS %v", m))
 	c := new(dns.Client)
 	m.SetTsig(keyName, dns.HmacMD5, 300, time.Now().Unix())
 	c.TsigSecret = map[string]string{keyName: tsigKey}
 	_, _, err := c.Exchange(m, dnsManager+":53")
 	if err != nil {
-		lbc.WriteToLog("ERROR", fmt.Sprintf("DNS update failed with (%v)", err))
+		lbc.WriteToLog(log.LevelError, fmt.Sprintf("DNS update failed with (%v)", err))
 		return err
 	}
-	lbc.WriteToLog("INFO", fmt.Sprintf("DNS update with keyName %v", keyName))
+	lbc.WriteToLog(log.LevelInfo, fmt.Sprintf("DNS update with keyName %v", keyName))
 
 	return nil
 }
@@ -85,7 +114,7 @@ func (lbc *LBCluster) getIpsFromDNS(m *dns.Msg, dnsManager string, dnsType uint1
 	m.SetQuestion(lbc.ClusterName+".", dnsType)
 	in, err := dns.Exchange(m, dnsManager+":53")
 	if err != nil {
-		lbc.WriteToLog("ERROR", fmt.Sprintf("Error getting the ipv4 state of dns: %v", err))
+		lbc.WriteToLog(log.LevelError, fmt.Sprintf("Error getting the ipv4 state of dns: %v", err))
 		return err
 	}
 	for _, a := range in.Answer {
@@ -104,7 +133,7 @@ func (lbc *LBCluster) GetStateDNS(dnsManager string) error {
 	m := new(dns.Msg)
 	var ips []net.IP
 	m.SetEdns0(4096, false)
-	lbc.WriteToLog("DEBUG", "Getting the ips from the DNS")
+	lbc.WriteToLog(log.LevelDebug, "Getting the ips from the DNS")
 	err := lbc.getIpsFromDNS(m, dnsManager, dns.TypeA, &ips)
 
 	if err != nil {
@@ -115,7 +144,7 @@ func (lbc *LBCluster) GetStateDNS(dnsManager string) error {
 		return err
 	}
 
-	lbc.WriteToLog("INFO", fmt.Sprintf("Let's keep the list of ips : %v", ips))
+	lbc.WriteToLog(log.LevelInfo, fmt.Sprintf("Let's keep the list of ips : %v", ips))
 	lbc.PreviousBestIpsDns = ips
 
 	return nil
